fix: stackoverflow questions
add No handler Found error
